+++ wemake_python_styleguide/presets/general.py
@@ -9,15 +9,12 @@ from wemake_python_styleguide.visitors.ast import (
     conditions,
     exceptions,
     functions,
+    imports,
     keywords,
     loops,
+    modules,
     naming,
     statements,
-)
-from wemake_python_styleguide.visitors.ast.imports import WrongImportVisitor
-from wemake_python_styleguide.visitors.ast.modules import (
-    EmptyModuleContentsVisitor,
-    MagicModuleFunctionsVisitor,
 )
 from wemake_python_styleguide.visitors.filenames.module import (
     WrongModuleNameVisitor,
@@ -48,7 +45,7 @@ GENERAL_PRESET = (
     exceptions.WrongTryExceptVisitor,
     exceptions.NestedTryBlocksVisitor,
 
-    WrongImportVisitor,
+    imports.WrongImportVisitor,
 
     naming.WrongNameVisitor,
     naming.WrongModuleMetadataVisitor,
@@ -73,6 +70,7 @@ GENERAL_PRESET = (
 
     # Modules:
     WrongModuleNameVisitor,
-    EmptyModuleContentsVisitor,
-    MagicModuleFunctionsVisitor,
+    modules.EmptyModuleContentsVisitor,
+    modules.MagicModuleFunctionsVisitor,
+    modules.ModuleConstantsVisitor,
 )